Move file reading functionality to file reading module
