+++ lib/encrypt.rb
@@ -21,20 +21,8 @@ class Encrypt
     DateGenerator.generate
   end
 
-  def read_file
-    file = File.open(@filename)
-    file_data = file.readlines.map(&:chomp)
-    @phrase = file_data[0].downcase
-    @key = file_data[1]
-    @date = file_data[2]
-  end
-
   def encrypt
     @output = @enig.encrypt(@phrase, @key, @date)
-  end
-
-  def write_to_file
-    File.open(@outfilename, 'w') {|file| file.write @output}
   end
 
   def self.run
